second update with less html
I was using the <details> tag because it was in the official github documentation on markdown but technically that is not markdown. Thoughts? I promise, I am not trying to be contrary! But I almost left the tags in without realizing they were NOT markdown since I got that from the github documentation so I am just curious.